requirements.txt
django
requests
pydantic
tzdata
python-dotenv
